Change Patch Color on Heelstrike before
Unity should know when to change color. Need to figure out how to change
color, and fix the issue with the perturb status being set before
actually perturbing. Unity needs to see the perturb value on heelstrike
before perturbation, but for the purpose of data analysis, the perturb
status needs to be set on heelstrike of the perturbation